requirements.txt
Django
djangorestframework
django-cors-headers
tensorflow-cpu
numpy
pandas
gunicorn
whitenoise